Do not generate python bytecode files
Python bytecode files (`.pyc`s) are meant to improve performance (no
need to re-generate bytecode). But in our case, give that they are
written to the volume which is mapped from the host, we end up with
root:root files that are annoying when deleting the whole directory/repo.

Since this is just a development repo, if the developer-user is ok with
those annoying root:root files, they can simply comment out that env
variable.